+++ src/main/webapp/components/nav-menu/index.js
@@ -1,0 +1,2 @@
+import NavMenu from './nav-menu'
+export default NavMenu
